oasp/oasp4j#358: improved java package structure (avoid split packages for later Java9 support), added missing @since tags
