+++ modules/cxf-client/src/main/java/io/oasp/module/cxf/common/impl/client/interceptor/PerformanceStartInterceptor.java
@@ -1,4 +1,4 @@
-package io.oasp.module.service.common.impl.cxf.interceptor;
+package io.oasp.module.cxf.common.impl.client.interceptor;
 
 import net.sf.mmm.util.date.api.TimeMeasure;
 
@@ -9,6 +9,8 @@ import org.apache.cxf.phase.Phase;
 
 /**
  * Implementation of {@link AbstractPhaseInterceptor} that logs the duration time of a service client invocation.
+ *
+ * @since 3.0.0
  */
 public class PerformanceStartInterceptor extends AbstractPhaseInterceptor<Message> {
