+++ modules/rest/src/main/java/io/oasp/module/rest/common/api/RestService.java
@@ -8,8 +8,10 @@ import io.oasp.module.service.common.api.Service;
 
 /**
  * This is a marker interface for a REST {@link Service}. It is recommended to extend your REST API interfaces from this
- * {@link RestService} interface to make your life easier. However, you are not forced to do so. See JavaDoc of {@link Service} for further
- * details.
+ * {@link RestService} interface to make your life easier. However, you are not forced to do so. See JavaDoc of
+ * {@link Service} for further details.
+ *
+ * @since 3.0.0
  */
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
