+++ modules/service/src/main/java/io/oasp/module/service/common/api/header/ServiceHeaderCustomizer.java
@@ -5,7 +5,10 @@ package io.oasp.module.service.common.api.header;
  * {@link io.oasp.module.service.common.api.ServiceClientFactory#create(Class) service client} is invoked these headers
  * are applied to the underlying protocol when a remote invocation is triggered via the network. Multiple
  * implementations of this interface may exist as spring beans for different aspects (e.g. passing a JWT/OAuth security
- * token for authentication, passing
+ * token for authentication, passing {@link io.oasp.module.logging.common.api.LoggingConstants#CORRELATION_ID
+ * correlation ID}, etc.).
+ *
+ * @since 3.0.0
  */
 public interface ServiceHeaderCustomizer {
